requirements.txt
azure-identity
azure-monitor-opentelemetry
azure-search-documents
azure-storage-blob
fastapi
firecrawl-py
langchain
langchain-community 
langchain-openai
langgraph
langsmith
opentelemetry-instrumentation-fastapi
pandas
psycopg2-binary
pydantic
pypdf
python-dotenv
redis
requests
sqlalchemy
streamlit
uvicorn
yt-dlp